Moved The method into GenericRepresenataion File
